+++ crates/wdl-analysis/src/analyzer.rs
@@ -22,6 +22,7 @@ use line_index::LineIndex;
 use line_index::WideEncoding;
 use line_index::WideLineCol;
 use lsp_types::GotoDefinitionResponse;
+use lsp_types::Location;
 use path_clean::clean;
 use tokio::runtime::Handle;
 use tokio::sync::mpsc;
@@ -44,6 +45,7 @@ use crate::graph::ParseState;
 use crate::queue::AddRequest;
 use crate::queue::AnalysisQueue;
 use crate::queue::AnalyzeRequest;
+use crate::queue::FindAllReferencesRequest;
 use crate::queue::FormatRequest;
 use crate::queue::GotoDefinitionRequest;
 use crate::queue::NotifyChangeRequest;
@@ -747,6 +749,38 @@ where
             anyhow!(
                 "failed to receive goto definition response from analysis queue because the \
                  channel has closed"
+            )
+        })
+    }
+
+    /// Performs a `find references` for a symbol across all the documents.
+    pub async fn find_all_references(
+        &self,
+        document: Url,
+        position: SourcePosition,
+        encoding: SourcePositionEncoding,
+        include_declaration: bool,
+    ) -> Result<Vec<Location>> {
+        let (tx, rx) = oneshot::channel();
+        self.sender
+            .send(Request::FindAllReferences(FindAllReferencesRequest {
+                document,
+                position,
+                encoding,
+                include_declaration,
+                completed: tx,
+            }))
+            .map_err(|_| {
+                anyhow!(
+                    "failed to send find all references request to analysis queue because the \
+                     channel has closed"
+                )
+            })?;
+
+        rx.await.map_err(|_| {
+            anyhow!(
+                "failed to receive find all references response from analysis queue because the \
+                 client channel has closed"
             )
         })
     }