+++ crates/wdl-analysis/src/handlers.rs
@@ -1,5 +1,7 @@
 //! Language server protocol handlers.
 
+mod find_all_references;
 mod goto_definition;
 
+pub use find_all_references::*;
 pub use goto_definition::*;